requirements.txt
pytest==8.3.3
python-fasthtml==0.10.0